+++ packages/terra-abstract-modal/src/_ModalContent.jsx
@@ -138,15 +138,29 @@ const ModalContent = forwardRef((props, ref) => {
         ref={ref}
       >
         <FormattedMessage id="Terra.AbstractModal.BeginModalDialog">
-          {text => (
-            <VisuallyHiddenText data-terra-abstract-modal-begin tabIndex="-1" text={text} />
-          )}
+          {text => {
+            // In the latest version of react-intl this param is an array, when previous versions it was a string.
+            let useText = text;
+            if (Array.isArray(text)) {
+              useText = text.join('');
+            }
+            return (
+              <VisuallyHiddenText data-terra-abstract-modal-begin tabIndex="-1" text={useText} />
+            );
+          }}
         </FormattedMessage>
         {children}
         <FormattedMessage id="Terra.AbstractModal.EndModalDialog">
-          {text => (
-            <VisuallyHiddenText text={text} />
-          )}
+          {text => {
+            // In the latest version of react-intl this param is an array, when previous versions it was a string.
+            let useText = text;
+            if (Array.isArray(text)) {
+              useText = text.join('');
+            }
+            return (
+              <VisuallyHiddenText text={useText} />
+            );
+          }}
         </FormattedMessage>
       </div>
     </React.Fragment>