+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Updated FormattedMessage child function to join if given an array.
+  
 * Fixed
  * Fixed broken links in documentation.